requirements.txt
pandas>=2.0.0
matplotlib>=3.7.0
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=7.4.0
